+++ include/OthelloAI.hpp
@@ -9,13 +9,13 @@ class ScoreList {
 public:
   ScoreList(BoardSeries::Position pos);
   BoardSeries::Position get_position() const;
-  void add_score(double score);
+  void add_score(float score);
   void show_score_list() const;
   bool operator>(const ScoreList &right)const;
-  double get_score() const;
+  float get_score() const;
 private:
   BoardSeries::Position pos_;
-  double score_;
+  float score_;
 };
 
 class OthelloAI {
@@ -26,7 +26,7 @@ public:
   ~OthelloAI();
   void set_score_list();
   BoardSeries::Position get_conclusion();
-  double get_sub_score();
+  float get_sub_score();
 private:
   BoardSeries::AiBoard myboard_;
   unsigned short branch_;
