+++ src/AiBoard.cpp
@@ -7,10 +7,10 @@ AiBoard::AiBoard(BoardBase src)
 {
 }
 
-double AiBoard::get_status_score() const
+float AiBoard::get_status_score() const
 { // とりあえず既存のやつ。gsliceもalgorithmもめっちゃ使えそう
   const Stone enemy_stone {get_enemy_stone()};
-  double status_score {static_cast<double>(count_stone(active_stone_) - count_stone(enemy_stone))};
+  float status_score {static_cast<float>(count_stone(active_stone_) - count_stone(enemy_stone))};
   Stone target{Stone::Space};
   static const std::vector<Position> edge_pos = {
     {0, 0}, {7, 0},
